--- fixtures/app_with_framework_and_tests_with_custom_filenames/App/AppDelegate.swift
@@ -1,22 +1,0 @@
-import Framework
-import UIKit
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-    var window: UIWindow?
-    var framework: FrameworkClass = FrameworkClass()
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-        window = UIWindow(frame: UIScreen.main.bounds)
-        let viewController = UIViewController()
-        viewController.view.backgroundColor = .white
-        window?.rootViewController = viewController
-        window?.makeKeyAndVisible()
-        return true
-    }
-    
-    func hello() -> String {
-        return "AppDelegate.hello()"
-    }
-    
-}
