--- fixtures/app_with_framework_and_tests_with_custom_filenames/AppTests/AppTests.swift
@@ -1,14 +1,0 @@
-import Foundation
-import XCTest
-
-@testable import App
-
-final class AppTests: XCTestCase {
-    
-    func testHello() {
-        let sut = AppDelegate()
-        
-        XCTAssertEqual("AppDelegate.hello()", sut.hello())
-    }
-    
-}
