--- fixtures/app_with_framework_and_tests_with_custom_filenames/Framework/Framework.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public class FrameworkClass {
-    public init() {}
-}
