--- fixtures/app_with_framework_and_tests_with_custom_filenames/FrameworkTests/FrameworkTests.swift
@@ -1,7 +1,0 @@
-import Foundation
-import XCTest
-@testable import Framework
-
-final class FrameworkTests: XCTestCase {
-    var subject: FrameworkClass = FrameworkClass()
-}
