--- fixtures/app_with_framework_and_tests_with_custom_filenames/TuistConfig.swift
@@ -1,8 +1,0 @@
-import ProjectDescription
-
-let config = TuistConfig(
-  generationOptions: [
-    .generateManifest,
-    .xcodeProjectName("AwesomePrefix-\(TemplateString.Token.projectName)-AwesomeSuffix")
-  ]
-)
